consumer_avformat.c: bugfix (kdenlive-28) a/v sync on non-whole frame rate.
git-svn-id: https://mlt.svn.sourceforge.net/svnroot/mlt/trunk/mlt@1050 d19143bc-622f-0410-bfdd-b5b2a6649095

+++ src/modules/avformat/consumer_avformat.c
@@ -623,7 +623,7 @@ static void *consumer_thread( void *arg )
 	struct timeval ante;
 
 	// Get the frame rate
-	int fps = mlt_properties_get_double( properties, "fps" );
+	double fps = mlt_properties_get_double( properties, "fps" );
 
 	// Get width and height
 	int width = mlt_properties_get_int( properties, "width" );
